Update the displayed release information output
This commit contains changes which updates the release information displayed
in the `opa run` and `opa version` commands.

In the server mode, log message has been updated for the OPA up-to-date and
out-of-date scenarios.

In the repl mode, release information is part of the `help` output.

For the `opa version` command, the displayed output is similar for both
OPA up-to-date and out-of-date scenarios.

Signed-off-by: Ashutosh Narkar <anarkar4387@gmail.com>