fixup: refactor to avoid mixing presentation logic
This commit just refactors the implementation a bit to avoid mixing
presentation and business logic. For instance, the REPL should decide
how to present the version report information in the help text--the
version report is structured data that can be passed around and
inspected. By converting the report to a string and passing that
around (and testing on that) consumers are forced to program against
strings which is generally not good--you end up with brittle code that
has to be modified frequently to account for small changes in
presentation. Generally, avoid converting data into strings unless the
consumer only cares about "strings" and not the information
represented by those strings. If the version report was not internal,
I would have probably just passed it directly into the REPL. We can
revisit this in the future.

The other thing this commit changed is the checkOPAVersion function in
the runtime. It was unclear why the done channel wa being passed into
that function so I just removed it and updated the callers.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>