wip: Spike out a text-based debugger using tracing
This is just a quick hack to see how much work it would be to standup
a simple text-based debugger on top of the topdown tracing
support

The debugger works by providing a custom tracer that lets users set
breakpoints on source lines and then enters a prompt and waits for
input when topdown evaluates (or redoes) expressions on any matching
lines. When a breakpoint is encountered, users can step to or goto the
next line (skipping function calls). Users can dump local variables to
see their values.

Many improvements could be made:

* Clean separation between debugger, prompting, and presentation
* Clean handling of input (e.g., current parsing is just a hack)
* Better support for evaluating variables when a breakpoint hits
* Better support for ad-hoc queries while debugging
* Support for rewinding
* Support for listing source code
* ...etc.