ci: Update post-tag workflow to decouple tag and asset dir
This change modifies the post-tag workflow to set the asset dir based
on the RELEASE_DIR rather than the tag name. The RELEASE_DIR is named
after the version that is hardcoded into version/version.go--since the
tag can be set to anything, the asset dir should use the RELEASE_DIR
rather than the tag name.

Signed-off-by: Torin Sandall <torinsandall@gmail.com>