wip: early-exit: wrap EE signal to support nested cases
see TODO in changes--we would be better off using a pointer to the
previous EE signal and to include the eval struct pointer in the EE
signal for traceability purposes.

conceptually that's the same solution though; just keep a stack of EE
signals and only suppress EE entirely if the stack is empty.