wip: skip recompilation if .rego files have not changed
This is an experimental change to see if we can reduce cpu usage
meaningfully by skipping recompilation when .rego files in snapshot
bundles have not changed.

The approach in this commit is to:
1. Add checksum to modules produced by the parser
2. Compare checksums in the bundle activation path
3. Store checksums on the compiler object

This seemed like the cleanest option for implementing this kind of
optimization because of the set of modules that are passed into the
compiler is not finalized until deep into the bundle/store.go activation
code path. Moreoever, the storing the checksums on the compiler seemed
natural since the bundle plugin can easily grab the old checksums off
the compiler via the plugin manager.